+++ waveq3d/eigenverb_model.h
@@ -52,10 +52,10 @@ class USML_DECLSPEC eigenverb_model : public reverberation_model {
          * @param wave          Wave queue, used to extract various data
          * @param ID            (Used to identify source/receiver/volume layer)
          */
-        virtual void notifyUpperCollision( unsigned de, unsigned az,
+        virtual void notifyUpperCollision( size_t de, size_t az,
                double dt, double grazing, double speed,
                const wposition1& position, const wvector1& ndirection,
-               const wave_queue& wave, unsigned ID ) {}
+               const wave_queue& wave, size_t ID ) {}
 
         /**
          * React to the collision of a single ray with a reverberation
@@ -71,10 +71,10 @@ class USML_DECLSPEC eigenverb_model : public reverberation_model {
          * @param wave          Wave queue, used to extract various data
          * @param ID            (Used to identify source/receiver/volume layer)
          */
-        virtual void notifyLowerCollision( unsigned de, unsigned az,
+        virtual void notifyLowerCollision( size_t de, size_t az,
                double dt, double grazing, double speed,
                const wposition1& position, const wvector1& ndirection,
-               const wave_queue& wave, unsigned ID ) {}
+               const wave_queue& wave, size_t ID ) {}
 
         /**
          * Computes the reverberation curve from the data cataloged from the
@@ -117,7 +117,7 @@ class USML_DECLSPEC eigenverb_model : public reverberation_model {
          * @param wave          Wave queue, used to extract various data
          * @return verb         newly constructed eigenverb
          */
-        void create_eigenverb( unsigned de, unsigned az,
+        void create_eigenverb( size_t de, size_t az,
                double dt, double grazing, double speed,
                const wposition1& position, const wvector1& ndirection,
                const wave_queue& wave, eigenverb& verb ) const ;
@@ -160,7 +160,7 @@ class USML_DECLSPEC eigenverb_model : public reverberation_model {
         /**
          * Number of time bins to store the energy distribution
          */
-        unsigned _max_index ;
+        size_t _max_index ;
 
         /**
          * Max time for the reverberation curve
@@ -170,17 +170,17 @@ class USML_DECLSPEC eigenverb_model : public reverberation_model {
         /**
          * Number of layers within the volume
          */
-        unsigned _n ;
+        size_t _n ;
 
         /**
          * Origin ID of the source wavefront
          */
-        unsigned _source_origin ;
+        size_t _source_origin ;
 
         /**
          * Origin ID of the reciever wavefront
          */
-        unsigned _receiver_origin ;
+        size_t _receiver_origin ;
 
         /**
          * Defines the type of boundary model for the bottom.
